reorganize entities folder and files
