+++ payment/src/main/java/com/rider/payment/entities/invoice/InvoiceType.java
@@ -1,4 +1,4 @@
-package com.rider.payment.entities.invoices;
+package com.rider.payment.entities.invoice;
 
 public enum InvoiceType {
     PAYMENT, REFUND
