+++ payment/src/main/java/com/rider/payment/entities/payment/PaymentStatus.java
@@ -1,4 +1,4 @@
-package com.rider.payment.entities.invoices;
+package com.rider.payment.entities.payment;
 
 public enum PaymentStatus {
     PENDING, FAILED, SUCCESSFUL, REFUNDED
