+++ payment/src/main/java/com/rider/payment/entities/paymentMethods/PaymentType.java
@@ -1,4 +1,4 @@
-package com.rider.payment.entities.PaymentMethod;
+package com.rider.payment.entities.paymentMethods;
 
 public enum PaymentType {
     CREDIT_CARD, PIX
